+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Put unreleased changes here!
 Builtin Apps:
 - About: fix logo not showing
 - AppStore: download the correct .mpk from BadgeHub.eu if there are multiple versions
+- OSUpdate: show raw ESP error and a user-friendly message when the update fails to activate
 
 Frameworks:
 - WifiService: increase connection timeout from 10 to 13 seconds